pyproject.toml
[project]
name = "image-processing-service"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.3",
    "uvicorn>=0.43.0",
    "pillow",
    "boto3",
    "sqlalchemy",
    "alembic",
    "pyjwt",
    "bcrypt",
    "redis",
    "pydantic[email]",
    "pydantic-settings",
    "python-multipart>=0.0.24",
    "asyncpg",
]

[dependency-groups]
dev = [
    "hypothesis",
    "pytest",
    "pytest-asyncio",
    "httpx",
    "aiosqlite>=0.22.1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.hypothesis]
max_examples = 100
